+++ unittests/tuple.cpp
@@ -24,96 +24,29 @@
   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#include "common.hpp"
-#include <boost/test/unit_test.hpp>
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include "tuple.hpp"
 
-template <class IArchive, class OArchive>
-void test_tuple()
-{
-  std::random_device rd;
-  std::mt19937 gen(rd());
+TEST_SUITE("tuple");
 
-  auto rng = [&](){ return random_value<int>(gen); };
-
-  for(int ii=0; ii<100; ++ii)
-  {
-    auto o_podtuple = std::make_tuple( rng(), rng(), rng(), rng() );
-    auto o_podtuple11 = std::make_tuple( rng(), rng(), rng(), rng(), rng(), rng(),
-                                         rng(), rng(), rng(), rng(), rng() );
-    auto o_isertuple = std::make_tuple( StructInternalSerialize( rng(), rng() ),
-        StructInternalSerialize( rng(), rng() ),
-        StructInternalSerialize( rng(), rng() ),
-        StructInternalSerialize( rng(), rng() ) );
-    auto o_ispltuple = std::make_tuple( StructInternalSplit( rng(), rng() ),
-        StructInternalSplit( rng(), rng() ),
-        StructInternalSplit( rng(), rng() ),
-        StructInternalSplit( rng(), rng() ) );
-    auto o_esertuple = std::make_tuple( StructExternalSerialize( rng(), rng() ),
-        StructExternalSerialize( rng(), rng() ),
-        StructExternalSerialize( rng(), rng() ),
-        StructExternalSerialize( rng(), rng() ) );
-    auto o_espltuple = std::make_tuple( StructExternalSerialize( rng(), rng() ),
-        StructExternalSerialize( rng(), rng() ),
-        StructExternalSerialize( rng(), rng() ),
-        StructExternalSerialize( rng(), rng() ) );
-
-    std::ostringstream os;
-    {
-      OArchive oar(os);
-
-      oar(o_podtuple);
-      oar(o_podtuple11);
-      oar(o_isertuple);
-      oar(o_ispltuple);
-      oar(o_esertuple);
-      oar(o_espltuple);
-    }
-
-    decltype( o_podtuple   ) i_podtuple;
-    decltype( o_podtuple11 ) i_podtuple11;
-    decltype( o_isertuple  ) i_isertuple;
-    decltype( o_ispltuple  ) i_ispltuple;
-    decltype( o_esertuple  ) i_esertuple;
-    decltype( o_espltuple  ) i_espltuple;
-
-    std::istringstream is(os.str());
-    {
-      IArchive iar(is);
-
-      iar(i_podtuple);
-      iar(i_podtuple11);
-      iar(i_isertuple);
-      iar(i_ispltuple);
-      iar(i_esertuple);
-      iar(i_espltuple);
-    }
-
-    BOOST_CHECK_EQUAL( i_podtuple   == o_podtuple,   true );
-    BOOST_CHECK_EQUAL( i_podtuple11 == o_podtuple11, true );
-    BOOST_CHECK_EQUAL( i_isertuple  == o_isertuple,  true );
-    BOOST_CHECK_EQUAL( i_ispltuple  == o_ispltuple,  true );
-    BOOST_CHECK_EQUAL( i_esertuple  == o_esertuple,  true );
-    BOOST_CHECK_EQUAL( i_espltuple  == o_espltuple,  true );
-  }
-}
-
-BOOST_AUTO_TEST_CASE( binary_tuple )
+TEST_CASE("binary_tuple")
 {
   test_tuple<cereal::BinaryInputArchive, cereal::BinaryOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( portable_binary_tuple )
+TEST_CASE("portable_binary_tuple")
 {
   test_tuple<cereal::PortableBinaryInputArchive, cereal::PortableBinaryOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( xml_tuple )
+TEST_CASE("xml_tuple")
 {
   test_tuple<cereal::XMLInputArchive, cereal::XMLOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( json_tuple )
+TEST_CASE("json_tuple")
 {
   test_tuple<cereal::JSONInputArchive, cereal::JSONOutputArchive>();
 }
 
+TEST_SUITE_END();
