+++ unittests/unordered_loads.cpp
@@ -24,134 +24,19 @@
   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#include "common.hpp"
-#include <boost/test/unit_test.hpp>
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include "unordered_loads.hpp"
 
-struct unordered_naming
-{
-  int x;
-  int xx;
-  int y;
-  int z;
+TEST_SUITE("unordered_loads");
 
-  template <class Archive>
-  void save( Archive & ar ) const
-  {
-    ar( CEREAL_NVP(x),
-        CEREAL_NVP(z),
-        CEREAL_NVP(y),
-        CEREAL_NVP(xx) );
-  }
-
-  template <class Archive>
-  void load( Archive & ar )
-  {
-    ar( x,
-        CEREAL_NVP(y),
-        CEREAL_NVP(z),
-        CEREAL_NVP(xx) );
-  }
-
-  bool operator==( unordered_naming const & other ) const
-  {
-    return x == other.x && xx == other.xx && y == other.y && z == other.z;
-  }
-};
-
-std::ostream& operator<<(std::ostream& os, unordered_naming const & s)
-{
-  os << "[x: " << s.x << " xx: " << s.xx << " y: " << s.y << " z: " << s.z << "]";
-  return os;
-}
-
-template <class IArchive, class OArchive>
-void test_unordered_loads()
-{
-  std::random_device rd;
-  std::mt19937 gen(rd());
-
-  auto rngB = [&](){ return random_value<int>( gen ) % 2 == 0; };
-  auto rngI = [&](){ return random_value<int>( gen ); };
-  auto rngF = [&](){ return random_value<float>( gen ); };
-  auto rngD = [&](){ return random_value<double>( gen ); };
-  auto rngS = [&](){ return random_basic_string<char>( gen ); };
-
-  for(int ii=0; ii<100; ++ii)
-  {
-    auto const name1 = rngS();
-    auto const name2 = rngS();
-    auto const name3 = rngS();
-    auto const name4 = rngS();
-    auto const name5 = rngS();
-    auto const name6 = rngS();
-    auto const name7 = rngS();
-
-    int o_int1 = rngI();
-    double o_double2 = rngD();
-    std::vector<bool> o_vecbool3 = { rngB(), rngB(), rngB(), rngB(), rngB() };
-    int o_int4 = rngI();
-    int o_int5 = rngI();
-    int o_int6 = rngI();
-    std::pair<float, unordered_naming> o_un7;
-    o_un7.first = rngF();
-    o_un7.second.x = rngI();
-    o_un7.second.xx = rngI();
-    o_un7.second.y = rngI();
-    o_un7.second.z = rngI();
-
-    std::ostringstream os;
-    {
-      OArchive oar(os);
-
-      oar( cereal::make_nvp( name1, o_int1 ),
-           cereal::make_nvp( name2, o_double2 ),
-           cereal::make_nvp( name3, o_vecbool3 ),
-           cereal::make_nvp( name4, o_int4 ),
-           cereal::make_nvp( name5, o_int5 ),
-           cereal::make_nvp( name6, o_int6 ),
-           cereal::make_nvp( name7, o_un7 ) );
-    }
-
-    decltype(o_int1) i_int1;
-    decltype(o_double2) i_double2;
-    decltype(o_vecbool3) i_vecbool3;
-    decltype(o_int4) i_int4;
-    decltype(o_int5) i_int5;
-    decltype(o_int6) i_int6;
-    decltype(o_un7) i_un7;
-
-    std::istringstream is(os.str());
-    {
-      IArchive iar(is);
-
-      iar( cereal::make_nvp( name7, i_un7 ),
-           cereal::make_nvp( name2, i_double2 ),
-           cereal::make_nvp( name4, i_int4 ),
-           cereal::make_nvp( name3, i_vecbool3 ),
-           cereal::make_nvp( name1, i_int1 ),
-           cereal::make_nvp( name5, i_int5 ),
-           i_int6 );
-    }
-
-    BOOST_CHECK_EQUAL(o_int1, i_int1);
-    BOOST_CHECK_CLOSE(o_double2 , o_double2, 1e-5);
-    BOOST_CHECK_EQUAL(o_vecbool3.size(), i_vecbool3.size());
-    BOOST_CHECK_EQUAL_COLLECTIONS(i_vecbool3.begin(),    i_vecbool3.end(),    o_vecbool3.begin(),  o_vecbool3.end());
-    BOOST_CHECK_EQUAL(o_int4, i_int4);
-    BOOST_CHECK_EQUAL(o_int5, i_int5);
-    BOOST_CHECK_EQUAL(o_int6, i_int6);
-    BOOST_CHECK_EQUAL(o_un7.first, i_un7.first);
-    BOOST_CHECK_EQUAL(o_un7.second, i_un7.second);
-  }
-}
-
-BOOST_AUTO_TEST_CASE( xml_unordered_loads )
+TEST_CASE("xml_unordered_loads")
 {
   test_unordered_loads<cereal::XMLInputArchive, cereal::XMLOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( json_unordered_loads )
+TEST_CASE("json_unordered_loads")
 {
   test_unordered_loads<cereal::JSONInputArchive, cereal::JSONOutputArchive>();
 }
 
+TEST_SUITE_END();
